Добавил валидацию и начал делать авторизацию, зарегистрировал сервисы

+++ aspnetPractice/Models/Client.cs
@@ -4,7 +4,7 @@ public class Client
     public int Id {get; set;}
     public string Name {get; set;} = "";
     public string Surname {get; set;} = "";
-    public short Age {get; set;}
+    public int Age {get; set;}
     public decimal Balance {get; set;}
 
 }
